Name change for tcpsink to be websocketsink
