+++ plugins/CMakeLists.txt
@@ -18,4 +18,4 @@ target_link_libraries(examplesinkplugin -lamb -L${CMAKE_CURRENT_BINARY_DIR}/lib 
 
 add_subdirectory(wheel)
 add_subdirectory(dbus)
-add_subdirectory(tcpsink)
+add_subdirectory(websocketsink)
